+++ packages/cli/__tests__/zosjobs/__integration__/download/__scripts__/download-output/conflicting_options.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+zowe zos-jobs download output JOB1234 --binary --record
+exit $?